modify: make inventory level numeric

+++ kori/app/models/store_product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer
+from sqlalchemy import Column, ForeignKey, Integer, Numeric
 from sqlalchemy.orm import relationship
 
 from kori.app.db.base import Base
@@ -7,8 +7,8 @@ from kori.app.db.base import Base
 class StoreProduct(Base):
     product_id = Column(ForeignKey("product.id"), primary_key=True)
     store_id = Column(ForeignKey("store.id"), primary_key=True)
-    stock_available = Column(Integer)
-    stock_locked = Column(Integer)
+    stock_available = Column(Numeric)
+    stock_locked = Column(Numeric)
 
     store = relationship("Store", back_populates="store_products")
     product = relationship("Product", back_populates="store_products")
